main: move optscript related code in lregex.c to newly introduced script.c
