+++ main/lregex.c
@@ -42,6 +42,7 @@
 #include "read_p.h"
 #include "routines.h"
 #include "routines_p.h"
+#include "script_p.h"
 #include "trashbox.h"
 #include "xtag_p.h"
 
@@ -247,11 +248,11 @@ static bool   guestRequestIsFilled(struct guestRequest *);
 static void   guestRequestClear (struct guestRequest *);
 static void   guestRequestSubmit (struct guestRequest *);
 
-static void optscriptSetup (OptVM *vm, struct lregexControlBlock *lcb, int corkIndex);
-static EsObject* optscriptRun (OptVM *vm, EsObject *optscript);
-static void optscriptTeardown (OptVM *vm, struct lregexControlBlock *lcb);
-
-static void optscriptInstallProcs (EsObject *dict);
+static EsObject *scriptRead (OptVM *vm, const char *src);
+static void scriptSetup (OptVM *vm, struct lregexControlBlock *lcb, int corkIndex);
+static EsObject* scriptEval (OptVM *vm, EsObject *optscript);
+static void scriptEvalPrelude (OptVM *vm, EsObject *dict, ptrArray *prelude);
+static void scriptTeardown (OptVM *vm, struct lregexControlBlock *lcb);
 
 static void deleteTable (void *ptrn)
 {
@@ -1251,20 +1252,6 @@ static void setKind(regexPattern * ptrn, const langType owner,
 	}
 }
 
-static EsObject *makeOptscriptObject (OptVM *vm, const char *src)
-{
-	size_t len = strlen (src);
-	Assert (len > 2);
-	Assert (src[len - 1] == '}');
-	Assert (src[len - 2] == '}');
-	MIO *mio = mio_new_memory ((unsigned char *)src + 1, len - 1 -1, NULL, NULL);
-	EsObject *obj = opt_vm_read (vm, mio);
-	if (es_error_p (obj))
-		error (FATAL, "failed in loading an optscript: %s", src);
-	mio_unref (mio);
-	return obj;
-}
-
 static void patternEvalFlags (struct lregexControlBlock *lcb,
 							  regexPattern * ptrn,
 							  enum regexParserType regptype,
@@ -1282,7 +1269,7 @@ static void patternEvalFlags (struct lregexControlBlock *lcb,
 	const char * optscript = flagsEval (flags, commonSpecFlagDef, ARRAY_SIZE(commonSpecFlagDef), &commonFlagData);
 	if (optscript)
 	{
-		ptrn->optscript = makeOptscriptObject (optvm, optscript);
+		ptrn->optscript = scriptRead (optvm, optscript);
 		ptrn->optscript_src = eStrdup (optscript);
 	}
 
@@ -1663,12 +1650,12 @@ static void matchTagPattern (struct lregexControlBlock *lcb,
 	{
 		if (patbuf->optscript)
 		{
-			optscriptSetup (optvm, lcb, n);
-			EsObject *e = optscriptRun (optvm, patbuf->optscript);
+			scriptSetup (optvm, lcb, n);
+			EsObject *e = scriptEval (optvm, patbuf->optscript);
 			if (es_error_p (e))
 				error (WARNING, "error when evaluating: %s", patbuf->optscript_src);
 			es_object_unref (e);
-			optscriptTeardown (optvm, lcb);
+			scriptTeardown (optvm, lcb);
 		}
 	}
 
@@ -1784,42 +1771,6 @@ static bool fillGuestRequest (const char *start,
 	return guestRequestIsFilled (guest_req);
 }
 
-static EsObject* optscriptRun (OptVM *vm, EsObject *optscript)
-{
-	static EsObject *exec = es_nil;
-
-	if (es_null (exec))
-	{
-		MIO *mio = mio_new_memory ((unsigned char*)"//exec", 6, NULL, NULL);
-		exec = opt_vm_read (vm, mio);
-		if (es_error_p (exec))
-			error (FATAL, "failed in converting //exec to an optscript object");
-		mio_unref (mio);
-	}
-
-	EsObject *o = opt_vm_eval (vm, optscript);
-	if (es_error_p (o))
-		error (FATAL, "failed to push the proc representing the script");
-	es_object_unref (o);
-
-	return opt_vm_eval (vm, exec);
-}
-
-static void optscriptRunPrelude (OptVM *vm, EsObject *dict, ptrArray *prelude)
-{
-	opt_vm_dstack_push (vm, dict);
-	for (int i = 0; i < ptrArrayCount (prelude); i++)
-	{
-		const char *src = ptrArrayItem (prelude, i);
-		EsObject * code = makeOptscriptObject (vm, src);
-		EsObject * e = optscriptRun (vm, code);
-		if (es_error_p (e))
-			error (WARNING, "error when evaluating: %s", src);
-		es_object_unref (code);
-	}
-	opt_vm_dstack_pop (vm);
-}
-
 static bool matchRegexPattern (struct lregexControlBlock *lcb,
 							   const vString* const line,
 							   regexTableEntry *entry)
@@ -1841,12 +1792,12 @@ static bool matchRegexPattern (struct lregexControlBlock *lcb,
 		entry->statistics.match++;
 		if (patbuf->optscript && (! hasNameSlot (patbuf)))
 		{
-			optscriptSetup (optvm, lcb, CORK_NIL);
-			EsObject *e = optscriptRun (optvm, patbuf->optscript);
+			scriptSetup (optvm, lcb, CORK_NIL);
+			EsObject *e = scriptEval (optvm, patbuf->optscript);
 			if (es_error_p (e))
 				error (WARNING, "error when evaluating: %s", patbuf->optscript_src);
 			es_object_unref (e);
-			optscriptTeardown (optvm, lcb);
+			scriptTeardown (optvm, lcb);
 		}
 
 		if (hasMessage(patbuf))
@@ -1924,12 +1875,12 @@ static bool matchMultilineRegexPattern (struct lregexControlBlock *lcb,
 		entry->statistics.match++;
 		if (patbuf->optscript && (! hasNameSlot (patbuf)))
 		{
-			optscriptSetup (optvm, lcb, CORK_NIL);
-			EsObject *e = optscriptRun (optvm, patbuf->optscript);
+			scriptSetup (optvm, lcb, CORK_NIL);
+			EsObject *e = scriptEval (optvm, patbuf->optscript);
 			if (es_error_p (e))
 				error (WARNING, "error when evaluating: %s", patbuf->optscript_src);
 			es_object_unref (e);
-			optscriptTeardown (optvm, lcb);
+			scriptTeardown (optvm, lcb);
 		}
 
 		if (patbuf->type == PTRN_TAG)
@@ -2021,7 +1972,7 @@ extern void notifyRegexInputStart (struct lregexControlBlock *lcb)
 	if (es_null (lcb->local_dict))
 	{
 		lcb->local_dict = opt_dict_new (23);
-		optscriptRunPrelude (optvm, lcb->local_dict, lcb->prelude);
+		scriptEvalPrelude (optvm, lcb->local_dict, lcb->prelude);
 	}
 	opt_vm_dstack_push (optvm, lcb->local_dict);
 }
@@ -2647,12 +2598,12 @@ static struct regexTable * matchMultitableRegexTable (struct lregexControlBlock 
 			entry->statistics.match++;
 			if (ptrn->optscript && (! hasNameSlot (ptrn)))
 			{
-				optscriptSetup (optvm, lcb, CORK_NIL);
-				EsObject *e = optscriptRun (optvm, ptrn->optscript);
+				scriptSetup (optvm, lcb, CORK_NIL);
+				EsObject *e = scriptEval (optvm, ptrn->optscript);
 				if (es_error_p (e))
 					error (WARNING, "error when evaluating: %s", ptrn->optscript_src);
 				es_object_unref (e);
-				optscriptTeardown (optvm, lcb);
+				scriptTeardown (optvm, lcb);
 			}
 
 			if (ptrn->type == PTRN_TAG)
@@ -2959,229 +2910,54 @@ static void   guestRequestSubmit (struct guestRequest *r)
 											r->boundary[BOUNDARY_END].offset);
 }
 
-static EsObject *optscript_CorkIndex_sym = es_nil;
+/*
+ * Script related functions
+ */
 
-static void optscriptSetup (OptVM *vm, struct lregexControlBlock *lcb, int corkIndex)
+/* This functions expects { code }} as input.
+ * Be care that curly brackets must be unbalanced.
+ */
+static EsObject *scriptRead (OptVM *vm, const char *src)
 {
-	if (corkIndex != CORK_NIL)
-	{
-		static EsObject *corkIndex_sym = es_nil;
-		if (es_null (corkIndex_sym))
-			corkIndex_sym = es_symbol_intern (".");
-		EsObject *corkIndex_val = es_integer_new (corkIndex);
-		opt_dict_def (lcb->local_dict, corkIndex_sym, corkIndex_val);
-		es_object_unref (corkIndex_val);
-		optscript_CorkIndex_sym = corkIndex_sym;
-	}
+	size_t len = strlen (src);
+	Assert (len > 2);
+	Assert (src[len - 1] == '}');
+	Assert (src[len - 2] == '}');
+
+	EsObject *obj = optscriptRead (vm, src + 1, len - 1 - 1);
+	if (es_error_p (obj))
+		error (FATAL, "failed in loading an optscript: %s", src);
+	return obj;
 }
 
-static void optscriptTeardown (OptVM *vm, struct lregexControlBlock *lcb)
+extern EsObject* scriptEval (OptVM *vm, EsObject *optscript)
 {
-	if (!es_null (optscript_CorkIndex_sym))
-	{
-		opt_dict_undef (lcb->local_dict, optscript_CorkIndex_sym);
-		optscript_CorkIndex_sym = es_nil;
-	}
+	return optscriptEval (vm, optscript);
 }
 
-static EsObject* lrop_get_field_value (OptVM *vm, EsObject *name)
+static void scriptEvalPrelude (OptVM *vm, EsObject *dict, ptrArray *prelude)
 {
-	EsObject *nobj = opt_vm_ostack_top (vm);
-	if (!es_integer_p (nobj))
-		return OPT_ERR_TYPECHECK;
-
-	int n = es_integer_get (nobj);
-	tagEntryInfo *e = getEntryInCorkQueue (n);
-	if (e == NULL)
+	opt_vm_dstack_push (vm, dict);
+	for (int i = 0; i < ptrArrayCount (prelude); i++)
 	{
-		EsObject *err = es_error_intern ("undefined");
-		return err;
+		const char *src = ptrArrayItem (prelude, i);
+		EsObject * code = scriptRead (vm, src);
+		EsObject * e = optscriptEval (vm, code);
+		if (es_error_p (e))
+			error (WARNING, "error when evaluating: %s", src);
+		es_object_unref (code);
 	}
-
-	void * data = es_symbol_get_data (name);
-	fieldType ftype = HT_PTR_TO_INT (data);
-	EsObject *val = getFieldValue (ftype, e);
-	if (es_error_p (val))
-		return val;
-
-	opt_vm_ostack_pop (vm);
-
-	if (isFieldValueAvailableAlways (ftype))
-	{
-		opt_vm_ostack_push (vm, val);
-		es_object_unref (val);
-	}
-	else if (es_null (val))
-	{
-		opt_vm_ostack_push (vm, es_false);
-	}
-	else
-	{
-		opt_vm_ostack_push (vm, val);
-		opt_vm_ostack_push (vm, es_true);
-		es_object_unref (val);
-	}
-	return es_false;
+	opt_vm_dstack_pop (vm);
 }
 
-static EsObject* lrop_set_field_value (OptVM *vm, EsObject *name)
+static void scriptSetup (OptVM *vm, struct lregexControlBlock *lcb, int corkIndex)
 {
-	EsObject *indexobj = opt_vm_ostack_peek (vm, 1);
-	if (!es_integer_p (indexobj))
-		return OPT_ERR_TYPECHECK;
-
-	int n = es_integer_get (indexobj);
-	tagEntryInfo *e = getEntryInCorkQueue (n);
-	if (e == NULL)
-	{
-		EsObject *err = es_error_intern ("undefined");
-		return err;
-	}
-
-	void * data = es_symbol_get_data (name);
-	fieldType ftype = HT_PTR_TO_INT (data);
-	unsigned int fdata_type = getFieldDataType (ftype);
-
-	EsObject *valobj = opt_vm_ostack_top (vm);
-	int valtype = es_object_get_type (valobj);
-
-	if (hasFieldValueCheckerForSetter (ftype))
-	{
-		EsObject *e = checkFieldValueForSetter (ftype, valobj);
-		if (!es_object_equal (e, es_false))
-			return e;
-	}
-	else
-	{
-		if (! (((fdata_type & FIELDTYPE_STRING) && (valtype == OPT_TYPE_STRING))
-			   || ((fdata_type & FIELDTYPE_BOOL) && (valtype == ES_TYPE_BOOLEAN))
-			   || ((fdata_type & FIELDTYPE_INTEGER) && (valtype == ES_TYPE_INTEGER))))
-			return OPT_ERR_TYPECHECK;
-	}
-
-	EsObject *r = setFieldValue (ftype, e, valobj);
-	if (es_error_p (r))
-		return r;
-
-	opt_vm_ostack_pop (vm);
-	opt_vm_ostack_pop (vm);
-
-	return es_false;
+	optscriptSetup (vm, lcb->local_dict, corkIndex);
 }
 
-static void optscriptInstallFieldGetter (EsObject *dict, fieldType ftype,
-										 vString *op_name, vString *op_desc)
+static void scriptTeardown (OptVM *vm, struct lregexControlBlock *lcb)
 {
-	const char *fname = getFieldName (ftype);
-	vStringPut (op_name, ':');
-	vStringCatS (op_name, fname);
-	EsObject *op_sym = es_symbol_intern (vStringValue (op_name));
-	es_symbol_set_data (op_sym, HT_INT_TO_PTR (ftype));
-
-	const char *vtype = getFieldGetterValueType (ftype);
-	unsigned int fdata_type = getFieldDataType (ftype);
-
-	vStringCatS (op_desc, "int :");
-	vStringCatS (op_desc, fname);
-	vStringPut (op_desc, ' ');
-
-	if (vtype)
-		vStringCatS (op_desc, vtype);
-	else
-	{
-		Assert (fdata_type);
-		if (fdata_type & FIELDTYPE_STRING)
-			vStringCatS (op_desc, "string|");
-		if (fdata_type & FIELDTYPE_INTEGER)
-			vStringCatS (op_desc, "int|");
-		if (fdata_type & FIELDTYPE_BOOL)
-			vStringCatS (op_desc, "bool|");
-		vStringChop (op_desc);
-	}
-
-	if (!isFieldValueAvailableAlways (ftype))
-	{
-		vStringPut (op_desc, ' ');
-		vStringCatS (op_desc, "true%");
-		vStringCatS (op_desc, "int :");
-		vStringCatS (op_desc, fname);
-		vStringCatS (op_desc, " false");
-	}
-
-	EsObject *op = opt_operator_new (lrop_get_field_value,
-									 vStringValue (op_name),
-									 1, vStringValue (op_desc));
-	opt_dict_def (dict, op_sym, op);
-	es_object_unref (op);
-}
-
-static void optscriptInstallFieldSetter (EsObject *dict, fieldType ftype,
-										 vString *op_name, vString *op_desc)
-{
-	const char *fname = getFieldName (ftype);
-	vStringCatS (op_name, fname);
-	vStringPut (op_name, ':');
-
-	EsObject *op_sym = es_symbol_intern (vStringValue (op_name));
-	es_symbol_set_data (op_sym, HT_INT_TO_PTR (ftype));
-
-	const char *vtype = getFieldSetterValueType (ftype);
-	unsigned int fdata_type = getFieldDataType (ftype);
-	vStringCatS (op_desc, "int ");
-
-	if (vtype)
-		vStringCatS (op_desc, vtype);
-	else
-	{
-		Assert (fdata_type);
-		if (fdata_type & FIELDTYPE_STRING)
-			vStringCatS (op_desc, "string|");
-		if (fdata_type & FIELDTYPE_INTEGER)
-			vStringCatS (op_desc, "int|");
-		if (fdata_type & FIELDTYPE_BOOL)
-			vStringCatS (op_desc, "bool|");
-		vStringChop (op_desc);
-	}
-
-	vStringPut (op_desc, ' ');
-	vStringCatS (op_desc, fname);
-	vStringCatS (op_desc, ": -");
-
-	EsObject *op = opt_operator_new (lrop_set_field_value,
-									 vStringValue (op_name),
-									 2, vStringValue (op_desc));
-	opt_dict_def (dict, op_sym, op);
-	es_object_unref (op);
-}
-
-static void optscriptInstallFieldAccessors (EsObject *dict)
-{
-	vString *op_name = vStringNew ();
-	vString *op_desc = vStringNew ();
-
-	for (fieldType ftype = 0; ftype <= FIELD_BUILTIN_LAST; ftype++)
-	{
-		if (hasFieldGetter (ftype))
-		{
-			optscriptInstallFieldGetter (dict, ftype, op_name, op_desc);
-			vStringClear (op_name);
-			vStringClear (op_desc);
-		}
-		if (hasFieldSetter (ftype))
-		{
-			optscriptInstallFieldSetter (dict, ftype, op_name, op_desc);
-			vStringClear (op_name);
-			vStringClear (op_desc);
-		}
-	}
-
-	vStringDelete (op_name);
-	vStringDelete (op_desc);
-}
-
-static void optscriptInstallProcs (EsObject *dict)
-{
-	optscriptInstallFieldAccessors (dict);
+	optscriptTeardown (vm, lcb->local_dict);
 }
 
 extern void	addOptscriptPrelude (struct lregexControlBlock *lcb, const char *code)
@@ -3208,17 +2984,7 @@ extern bool checkRegex (void)
 #endif
 
 	if (regexAvailable)
-	{
-		MIO *in  = mio_new_fp (stdin, NULL);
-		MIO *out = mio_new_fp (stdout, NULL);
-		MIO *err = mio_new_fp (stderr, NULL);
-
-		optvm = opt_vm_new (in, out, err);
-
-		mio_unref (err);
-		mio_unref (out);
-		mio_unref (in);
-	}
+		optvm = optscriptInit ();
 
 	return regexAvailable;
 }
